+++ HISTORY.md
@@ -1,4 +1,10 @@
-## [<ins>1.1.0 (15-08-2017)</ins>](https://github.com/tauquirahmed93/palette4bootstrap/releases/tag/v1.1.0)
+## [<ins>1.2.0 (28-08-2017)</ins>](https://github.com/tahmed1994/palette4bootstrap/releases/tag/v1.2.0)
+#### Added the following palette items
+* **Bootstrap Containers** (NEW)
+  * Container
+  * Panel
+
+## [<ins>1.1.0 (15-08-2017)</ins>](https://github.com/tahmed1994/palette4bootstrap/releases/tag/v1.1.0)
 #### Added the following palette items
 * **Bootstrap Typography**
   * More
@@ -10,7 +16,7 @@
   * Carousel
   * Google Maps
 
-## [<ins>1.0.0 (02-08-2017)</ins>](https://github.com/tauquirahmed93/palette4bootstrap/releases/tag/v1.0.0)
+## [<ins>1.0.0 (02-08-2017)</ins>](https://github.com/tahmed1994/palette4bootstrap/releases/tag/v1.0.0)
 ### Initial Commit
 #### Following palette items are available
 * **Bootstrap UI**